added code to load MNIST data to memory

+++ src/main/scala/Classifier.scala
@@ -1,11 +1,11 @@
 import breeze.linalg.{DenseVector, DenseMatrix}
 
 trait Classifier {
-  def train(data: DenseMatrix[Double], lables: DenseVector[Double])
-  def predict(data: DenseMatrix[Double]): DenseVector[Double]
+  def train(data: DenseMatrix[Double], lables: DenseVector[Int])
+  def predict(data: DenseMatrix[Double]): DenseVector[Int]
 }
 
 class LogistricRegression() extends Classifier {
-  override def train(data: DenseMatrix[Double], lables: DenseVector[Double]) = println(data.size)
-  override def predict(data: DenseMatrix[Double]): DenseVector[Double] = DenseVector.zeros[Double](data.size)
+  override def train(data: DenseMatrix[Double], lables: DenseVector[Int]) = println(data.size)
+  override def predict(data: DenseMatrix[Double]): DenseVector[Int] = DenseVector.zeros[Int](data.size)
 }
